+++ ps_funcs/py/__init__.py
@@ -20,6 +20,7 @@ from . import ps_csv
 from . import structs
 from . import strings
 from . import config
+from . import log