+++ ps_funcs/py/log.py
@@ -1,0 +1,6 @@
+"""
+This module contain functions for logging functionality.
+Most of logging parameters configurable through function arguments.
+"""
+import os, sys; import ps_funcs, ps_shcolar
+